Implémentation d'un certain nombre de fonctions de stdio.h, des FILE*, de stdout, stdin, stderr Ajustements sur les types des appels read, write (plus de size_t)

+++ kernel/include/kernel/filesystem.h
@@ -27,8 +27,8 @@ struct file_descr_t
 	int owners;
 	uint8_t flags;
 
-	size_t (*read)(struct file_descr_t*,void*,size_t);
-	size_t (*write)(struct file_descr_t*,void*,size_t);
+	int32_t (*read)(struct file_descr_t*,void*,int32_t);
+	int32_t (*write)(struct file_descr_t*,void*,int32_t);
 	uint32_t (*seek)(struct file_descr_t*,int32_t,int);
 	void (*close)(struct file_descr_t*);
 };
@@ -40,8 +40,8 @@ file_descr_t *open_inode_rd(uint32_t inode);
 file_descr_t *open_rd(const char *path);
 void close_rd(file_descr_t *fd);
 
-size_t read_rd(file_descr_t *fd, void *ptr, size_t count);
-size_t write_rd(file_descr_t *fd, void *ptr, size_t count);
+int32_t read_rd(file_descr_t *fd, void *ptr, int32_t count);
+int32_t write_rd(file_descr_t *fd, void *ptr, int32_t count);
 uint32_t seek_rd(file_descr_t *fd, int32_t ofs, int flag);
 
 #endif
